small refactoring to accommodate subscreens/components reusability
